compartmentalized some more
