+++ src/navigation/stacks/HealingResourcesNavigator.tsx
@@ -1,15 +1,13 @@
 import React from 'react';
-import { Text, TouchableOpacity, View } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
+import { Text, View } from 'react-native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import BackArrow from 'src/assets/images/back-arrow.svg';
 import Logo from 'src/assets/images/logo.svg';
 import { colors } from 'src/styles/colors';
+import BackButton from '@/components/BackButton';
 import HealingResources from '@/screens/HealingResources';
 import HealingCatalogue from '@/screens/HealingResources/HealingCatalogue';
 import HopeHealingGuide from '@/screens/HealingResources/HopeHealingGuide/';
 import { HealingStackParams } from '../types';
-import styles from './styles';
 
 const HealingStack = createNativeStackNavigator<HealingStackParams>();
 
@@ -51,7 +49,7 @@ export default function HealingResourcesNavigator() {
         options={{
           headerBackTitle: 'Healing Resources',
           headerTitle: '',
-          headerLeft: () => <BackButton />,
+          headerLeft: () => <BackButton label={'Healing Resources'} />,
           headerRight: () => (
             <View style={{ paddingRight: '2%' }}>
               <Logo />
@@ -65,7 +63,7 @@ export default function HealingResourcesNavigator() {
         options={{
           headerBackTitle: 'Healing Resources',
           headerTitle: '',
-          headerLeft: () => <BackButton />,
+          headerLeft: () => <BackButton label={'Healing Resources'} />,
           headerRight: () => (
             <View style={{ paddingRight: '2%' }}>
               <Logo />
@@ -75,17 +73,4 @@ export default function HealingResourcesNavigator() {
       />
     </HealingStack.Navigator>
   );
-  function BackButton() {
-    const navigation = useNavigation();
-
-    return (
-      <TouchableOpacity
-        onPress={() => navigation.goBack()}
-        style={styles.backContainer}
-      >
-        <BackArrow />
-        <Text style={styles.backText}>Healing Resources</Text>
-      </TouchableOpacity>
-    );
-  }
 }
