+++ src/navigation/stacks/LegalRightsNavigator.tsx
@@ -1,14 +1,12 @@
 import React from 'react';
-import { Text, TouchableOpacity, View } from 'react-native';
-import { useNavigation } from '@react-navigation/native';
+import { Text, View } from 'react-native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import BackArrow from 'src/assets/images/back-arrow.svg';
 import Logo from 'src/assets/images/logo.svg';
 import { colors } from 'src/styles/colors';
 import LegalRights from '@/screens/LegalRights';
 import VideoPage from '@/screens/LegalRights/VideoPage';
+import BackButton from '../../components/BackButton';
 import { LegalStackParams } from '../types';
-import styles from './styles';
 
 const LegalStack = createNativeStackNavigator<LegalStackParams>();
 
@@ -48,7 +46,7 @@ export default function LegalRightsNavigator() {
         options={{
           headerBackTitle: 'Legal Rights',
           headerTitle: '',
-          headerLeft: () => <BackButton />,
+          headerLeft: () => <BackButton label={'Legal Rights'} />,
           headerRight: () => (
             <View style={{ paddingRight: '2%' }}>
               <Logo />
@@ -58,17 +56,4 @@ export default function LegalRightsNavigator() {
       />
     </LegalStack.Navigator>
   );
-  function BackButton() {
-    const navigation = useNavigation();
-
-    return (
-      <TouchableOpacity
-        onPress={() => navigation.goBack()}
-        style={styles.backContainer}
-      >
-        <BackArrow />
-        <Text style={styles.backText}>Legal Rights</Text>
-      </TouchableOpacity>
-    );
-  }
 }
